Replace old site with modern Tailwind CSS site
- Move new site from content/newsite/ to content/
- Config-driven site generation with source/versions.json
- Templates in source/templates/ for index and downloads pages
- Run 'cd source && python3 generate.py' to regenerate
- Remove old markdown-based generation and CSS files
- Update BUILDING.txt and docs/README.md with new workflow
- Remove outdated blog/wiki links, update Twitter to X branding
- Update hero section to mention terabits/sec scale
- Update build commands to use CMake (ATS 10+)
