--- content/styles/1140.css
@@ -1,186 +1,0 @@
-/* CSS Resets */
-
-html,body,div,span,object,iframe,h1,h2,h3,h4,h5,h6,p,blockquote,pre,a,abbr,address,cite,code,del,dfn,em,img,ins,q,small,strong,sub,sup,dl,dt,dd,ol,ul,li,fieldset,form,label,legend,table,caption,tbody,tfoot,thead,tr,th,td{border:0;margin:0;padding:0}article,aside,figure,figure img,figcaption,hgroup,footer,header,nav,section,video,object{display:block}a img{border:0}figure{position:relative}figure img{width:100%}
-
-
-/* ==================================================================================================================== */
-/* ! The 1140px Grid V2 by Andy Taylor \ http://cssgrid.net \ http://www.twitter.com/andytlr \ http://www.andytlr.com   */
-/* ==================================================================================================================== */
-
-.container {
-padding-left: 20px;
-padding-right: 20px;
-}
-
-.warning {
-  margin-top: 1em;
-}
-
-.row .nav .eightcol {
-font-size: 11pt;
-}
-
-.row {
-width: 100%;
-max-width: 1140px;
-min-width: 755px;
-margin: 0 auto;
-overflow: hidden;
-}
-
-.onecol, .twocol, .threecol, .fourcol, .fivecol, .sixcol, .sevencol, .ninecol, .tencol, .elevencol {
-margin-right: 3.8%;
-float: left;
-min-height: 1px;
-}
-
-/* .eightcol used on front page causes weird whitespace at smaller windows */
-.eightcol{
-float: left;
-min-height: 1px;
-}
-
-
-.row .onecol {
-width: 4.85%;
-}
-
-.row .twocol {
-width: 13.45%;
-padding: 0.6em;
-}
-
-.row .threecol {
-width: 22.05%;
-}
-
-.row .fourcol {
-width: 30.75%;
-}
-
-.row .fivecol {
-width: 39.45%;
-}
-
-.row .sixcol {
-width: 48%;
-}
-
-.row .sevencol {
-width: 56.75%;
-}
-
-.row .eightcol {
-width: 65.4%;
-}
-
-.content > .eightcol {
-  width: 65.4%;
-}
-
-.row .ninecol {
-width: 74.05%;
-}
-
-.row .tencol {
-width: 82.7%;
-}
-
-.content > .tencol {
-  width: 100%;
-  margin-right: 0px;
-  padding: 1em;
-}
-
-.row .elevencol {
-width: 91.35%;
-}
-
-.row .twelvecol {
-width: 100%;
-float: left;
-}
-
-#footer > .tencol {
-  margin: 1em;
-}
-
-.last {
-margin-right: 0px;
-}
-
-img, object, embed {
-max-width: 100%;
-}
-
-img {
-	height: auto;
-}
-
-.comments {
-  padding: 20px;
-}
-
-.docsnav {
-  padding-top: 40px;
-  padding-left: 20px;
-}
-
-
-.eightcol table {
-  font-size: 12px;
-  background: #666;
-  padding: 0px;
-  margin: 0px;
-}
-.eightcol table th {
-  background: #F6F6FF;
-  padding: 3px;
-}
-.eightcol td {
-  background: #FFF;
-  padding: 3px;
-}
-
-/* Smaller screens */
-
-@media only screen and (max-width: 1023px) {
-
-	body {
-    /* Commented out two lines - leads to unpleasant text resizing, believe this is a mistake in the 1140.css*/
-	/*font-size: 0.8em;*/
-	/*line-height: 1.5em;*/
-	}
-
-	}
-
-
-/* Mobile */
-
-@media handheld, only screen and (max-width: 767px) {
-
-	body {
-    /* Commented out one line - leads to unpleasant text resizing, believe this is a mistake in the 1140.css*/
-    /*font-size: 16px;*/
-	-webkit-text-size-adjust: none;
-	}
-
-	.row, body, .container {
-	width: 100%;
-	min-width: 0;
-	margin-left: 0px;
-	margin-right: 0px;
-	padding-left: 0px;
-	padding-right: 0px;
-	}
-
-	.row .onecol, .row .twocol, .row .threecol, .row .fourcol, .row .fivecol, .row .sixcol, .row .sevencol, .row .eightcol, .row .ninecol, .row .tencol, .row .elevencol, .row .twelvecol {
-	width: auto;
-	float: none;
-	margin-left: 0px;
-	margin-right: 0px;
-	padding-left: 20px;
-	padding-right: 20px;
-	}
-
-}
