--- content/styles/ie.css
@@ -1,43 +1,0 @@
-.onecol {
-width: 4.7%;
-}
-
-.twocol {
-width: 13.2%;
-}
-
-.threecol {
-width: 22.05%;
-}
-
-.fourcol {
-width: 30.6%;
-}
-
-.fivecol {
-width: 39%;
-}
-
-.sixcol {
-width: 48%;
-}
-
-.sevencol {
-width: 56.75%;
-}
-
-.eightcol {
-width: 61.6%;
-}
-
-.ninecol {
-width: 74.05%;
-}
-
-.tencol {
-width: 82%;
-}
-
-.elevencol {
-width: 91.35%;
-}
