--- content/styles/pygments_style.css
@@ -1,62 +1,0 @@
-.codehilite .hll { background-color: #ffffcc }
-.codehilite  { background: #f0f0f0; }
-.codehilite .c { color: #60a0b0; font-style: italic } /* Comment */
-.codehilite .err { border: 1px solid #FF0000 } /* Error */
-.codehilite .k { color: #007020; font-weight: bold } /* Keyword */
-.codehilite .o { color: #666666 } /* Operator */
-.codehilite .cm { color: #60a0b0; font-style: italic } /* Comment.Multiline */
-.codehilite .cp { color: #007020 } /* Comment.Preproc */
-.codehilite .c1 { color: #60a0b0; font-style: italic } /* Comment.Single */
-.codehilite .cs { color: #60a0b0; background-color: #fff0f0 } /* Comment.Special */
-.codehilite .gd { color: #A00000 } /* Generic.Deleted */
-.codehilite .ge { font-style: italic } /* Generic.Emph */
-.codehilite .gr { color: #FF0000 } /* Generic.Error */
-.codehilite .gh { color: #000080; font-weight: bold } /* Generic.Heading */
-.codehilite .gi { color: #00A000 } /* Generic.Inserted */
-.codehilite .go { color: #808080 } /* Generic.Output */
-.codehilite .gp { color: #c65d09; font-weight: bold } /* Generic.Prompt */
-.codehilite .gs { font-weight: bold } /* Generic.Strong */
-.codehilite .gu { color: #800080; font-weight: bold } /* Generic.Subheading */
-.codehilite .gt { color: #0040D0 } /* Generic.Traceback */
-.codehilite .kc { color: #007020; font-weight: bold } /* Keyword.Constant */
-.codehilite .kd { color: #007020; font-weight: bold } /* Keyword.Declaration */
-.codehilite .kn { color: #007020; font-weight: bold } /* Keyword.Namespace */
-.codehilite .kp { color: #007020 } /* Keyword.Pseudo */
-.codehilite .kr { color: #007020; font-weight: bold } /* Keyword.Reserved */
-.codehilite .kt { color: #902000 } /* Keyword.Type */
-.codehilite .m { color: #40a070 } /* Literal.Number */
-.codehilite .s { color: #4070a0 } /* Literal.String */
-.codehilite .na { color: #4070a0 } /* Name.Attribute */
-.codehilite .nb { color: #007020 } /* Name.Builtin */
-.codehilite .nc { color: #0e84b5; font-weight: bold } /* Name.Class */
-.codehilite .no { color: #60add5 } /* Name.Constant */
-.codehilite .nd { color: #555555; font-weight: bold } /* Name.Decorator */
-.codehilite .ni { color: #d55537; font-weight: bold } /* Name.Entity */
-.codehilite .ne { color: #007020 } /* Name.Exception */
-.codehilite .nf { color: #06287e } /* Name.Function */
-.codehilite .nl { color: #002070; font-weight: bold } /* Name.Label */
-.codehilite .nn { color: #0e84b5; font-weight: bold } /* Name.Namespace */
-.codehilite .nt { color: #062873; font-weight: bold } /* Name.Tag */
-.codehilite .nv { color: #bb60d5 } /* Name.Variable */
-.codehilite .ow { color: #007020; font-weight: bold } /* Operator.Word */
-.codehilite .w { color: #bbbbbb } /* Text.Whitespace */
-.codehilite .mf { color: #40a070 } /* Literal.Number.Float */
-.codehilite .mh { color: #40a070 } /* Literal.Number.Hex */
-.codehilite .mi { color: #40a070 } /* Literal.Number.Integer */
-.codehilite .mo { color: #40a070 } /* Literal.Number.Oct */
-.codehilite .sb { color: #4070a0 } /* Literal.String.Backtick */
-.codehilite .sc { color: #4070a0 } /* Literal.String.Char */
-.codehilite .sd { color: #4070a0; font-style: italic } /* Literal.String.Doc */
-.codehilite .s2 { color: #4070a0 } /* Literal.String.Double */
-.codehilite .se { color: #4070a0; font-weight: bold } /* Literal.String.Escape */
-.codehilite .sh { color: #4070a0 } /* Literal.String.Heredoc */
-.codehilite .si { color: #70a0d0; font-style: italic } /* Literal.String.Interpol */
-.codehilite .sx { color: #c65d09 } /* Literal.String.Other */
-.codehilite .sr { color: #235388 } /* Literal.String.Regex */
-.codehilite .s1 { color: #4070a0 } /* Literal.String.Single */
-.codehilite .ss { color: #517918 } /* Literal.String.Symbol */
-.codehilite .bp { color: #007020 } /* Name.Builtin.Pseudo */
-.codehilite .vc { color: #bb60d5 } /* Name.Variable.Class */
-.codehilite .vg { color: #bb60d5 } /* Name.Variable.Global */
-.codehilite .vi { color: #bb60d5 } /* Name.Variable.Instance */
-.codehilite .il { color: #40a070 } /* Literal.Number.Integer.Long */
