--- source/markdown/acknowledgements.mdtext
@@ -1,28 +1,0 @@
-Title: Apache Traffic Server - Acknowledgements
-Notice:    Licensed to the Apache Software Foundation (ASF) under one
-           or more contributor license agreements.  See the NOTICE file
-           distributed with this work for additional information
-           regarding copyright ownership.  The ASF licenses this file
-           to you under the Apache License, Version 2.0 (the
-           "License"); you may not use this file except in compliance
-           with the License.  You may obtain a copy of the License at
-           .
-             http://www.apache.org/licenses/LICENSE-2.0
-           .
-           Unless required by applicable law or agreed to in writing,
-           software distributed under the License is distributed on an
-           "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-           KIND, either express or implied.  See the License for the
-           specific language governing permissions and limitations
-           under the License.
-		
-
-# Acknowledgements #
-
-The Apache Traffic Server team would like to acknowledge the support and 
-contributions of the following companies, organizations and individuals:
-
-* [Yahoo!](http://www.yahoo.com), for donating the original source code for the Apache Traffic Server to us.
-* [Rackspace](http://www.rackspace.com), for providing virtualized resources to our Continuous Integration system.
-* [GoDaddy](http://www.godaddy.com), for additional resources to our CI, as well as SSL certificates for our sites.
-* [The Measurement Factory](http://www.measurement-factory.com/), for letting us run their Co-Advisor test suite against Apache Traffic Server.
