--- source/markdown/assistance.mdtext
@@ -1,45 +1,0 @@
-Title: Third Party Assistance - Apache Traffic Server
-Notice:    Licensed to the Apache Software Foundation (ASF) under one
-           or more contributor license agreements.  See the NOTICE file
-           distributed with this work for additional information
-           regarding copyright ownership.  The ASF licenses this file
-           to you under the Apache License, Version 2.0 (the
-           "License"); you may not use this file except in compliance
-           with the License.  You may obtain a copy of the License at
-           .
-             https://www.apache.org/licenses/LICENSE-2.0
-           .
-           Unless required by applicable law or agreed to in writing,
-           software distributed under the License is distributed on an
-           "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-           KIND, either express or implied.  See the License for the
-           specific language governing permissions and limitations
-           under the License.
-
-# Third Party Assistance #
-This page is intended as an independently made list of companies offering 
-commercial or free support and consulting for the Apache Traffic Server. 
-
-This list is NOT an endorsement nor a recommendation by the Traffic Server 
-PMC, it is merely a list of companies and individuals who have expressed 
-that they offer support and consulting. For further information, please 
-contact the companies/individuals themselves - we do not nor will we ever 
-recommend one resource over the other.
-
-## Companies and individuals offering ATS support ##
-Sorted alphabetically
-
-[Mirimar Networks](http://www.mirimar.net/)
-
-[Network Geographics](https://www.network-geographics.com/) 
-
-[OmniTI](http://www.omniti.com/)  
-
-[We-Amp](http://www.we-amp.com/)
-
-Any company or individual that so wishes, can be put on this page by sending an 
-email to dev [at] trafficserver.apache.org, listing:
-
- * The name of the company/individual doing ATS support
- * A link to their/his/her web site
-
