--- source/markdown/downloads.mdtext
@@ -1,70 +1,0 @@
-Title: Apache Traffic Server Downloads
-Notice:    Licensed to the Apache Software Foundation (ASF) under one
-           or more contributor license agreements.  See the NOTICE file
-           distributed with this work for additional information
-           regarding copyright ownership.  The ASF licenses this file
-           to you under the Apache License, Version 2.0 (the
-           "License"); you may not use this file except in compliance
-           with the License.  You may obtain a copy of the License at
-           .
-             https://www.apache.org/licenses/LICENSE-2.0
-           .
-           Unless required by applicable law or agreed to in writing,
-           software distributed under the License is distributed on an
-           "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-           KIND, either express or implied.  See the License for the
-           specific language governing permissions and limitations
-           under the License.
-RSS:       /rss/releases.rss
-
-
-
-The latest stable release of Apache Traffic Server is v10.1.0, released on 2025-8-13.
-In addition, we continue to support the v9.2.x LTS release train, currently v9.2.10,
-which was released on 2025-4-02. We follow the [Semantic Versioning](http://semver.org)
-scheme. The goal is to release patch releases frequently, and minor releases as needed.
-Within the major versions, all such patch and minor releases are all compatible.
-
-When downloading from a mirror, please be sure to [verify the hash and OpenPGP
-compatible signature](https://www.apache.org/info/verification.html).  To validate
-the PGP signatures the [public keys from the release manager](https://www.apache.org/dist/trafficserver/KEYS)
-will be needed.  You can also
-[browse through all releases](https://archive.apache.org/dist/trafficserver/)
-and hash signatures.
-
-# Current v10.x Release -- 10.1.0 # {#10.1.0}
-
- Apache Traffic Server v10.1.0 was released on August 13th, 2025.
- [[`PGP`](https://www.apache.org/dist/trafficserver/trafficserver-10.1.0.tar.bz2.asc)]
- [[`SHA512`](https://www.apache.org/dist/trafficserver/trafficserver-10.1.0.tar.bz2.sha512)]
-
- <a href="https://www.apache.org/dyn/closer.cgi/trafficserver/trafficserver-10.1.0.tar.bz2"
- class="download_ts">Traffic Server 10.1.0</a>
-
-v10.0.6 is our latest stable release. Additional details for this release are in the
-[CHANGELOG](https://raw.githubusercontent.com/apache/trafficserver/10.1.x/CHANGELOG-10.1.0)
-and the the related [Github Issues and PRs](https://github.com/apache/trafficserver/pulls?q=is:closed+is:pr+milestone:10.1.0).
-
-For details on the v10.x release, please see
-[10.x News](https://docs.trafficserver.apache.org/en/latest/release-notes/whats-new.en.html). There are also
-details about [upgrading to 10.x](https://docs.trafficserver.apache.org/en/latest/release-notes/upgrading.en.html).
-
-# Current v9.x Release -- 9.2.11 # {#9.2.11}
-
- Apache Traffic Server v9.2.11 was released on June 17th, 2025.
- [[`PGP`](https://www.apache.org/dist/trafficserver/trafficserver-9.2.11.tar.bz2.asc)]
- [[`SHA512`](https://www.apache.org/dist/trafficserver/trafficserver-9.2.11.tar.bz2.sha512)]
-
- <a href="https://www.apache.org/dyn/closer.cgi/trafficserver/trafficserver-9.2.11.tar.bz2"
- class="download_ts">Traffic Server 9.2.11</a>
-
-For details on the v9.x release, please see
-[9.2.x News](https://docs.trafficserver.apache.org/en/9.2.x/release-notes/whats-new.en.html)
-(there's a new section specific to v9.2.x). There are also
-details about [upgrading to 9.x](https://docs.trafficserver.apache.org/en/9.2.x/release-notes/upgrading.en.html).
-
-# Archived Releases
-
-## [Apache Traffic Server Archive](https://archive.apache.org/dist/trafficserver/)
-
-<!-- test -->
