--- source/markdown/press.mdtext
@@ -1,44 +1,0 @@
-Title: Press Kit - Apache Traffic Server
-Notice:    Licensed to the Apache Software Foundation (ASF) under one
-           or more contributor license agreements.  See the NOTICE file
-           distributed with this work for additional information
-           regarding copyright ownership.  The ASF licenses this file
-           to you under the Apache License, Version 2.0 (the
-           "License"); you may not use this file except in compliance
-           with the License.  You may obtain a copy of the License at
-           .
-             https://www.apache.org/licenses/LICENSE-2.0
-           .
-           Unless required by applicable law or agreed to in writing,
-           software distributed under the License is distributed on an
-           "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-           KIND, either express or implied.  See the License for the
-           specific language governing permissions and limitations
-           under the License.
-
-# Press Kit #
-If you intend to use the Apache Traffic Server logo or icon for press releases, articles,
-or other publications, feel free to use any of the images below:
-
-
-![Traffic Server Logo](/images/press/ts_logo_tm_380x73.png "")
-
-[Traffic Server Logo, small](/images/press/ts_logo_tm_380x73.png)
-
-[Traffic Server Logo, medium](/images/press/ts_logo_tm_1024x196.png) 
-
-[Traffic Server Logo, large](/images/press/ts_logo_tm_1960x375.png) 
-
-[Traffic Server Logo, SVG](/images/press/trafficserver.svg)
-
-<br/>
-
-
-![Traffic Server Icon](/images/press/tsicon_small.png "")
-
-[Traffic Server Icon, Small](/images/press/tsicon_small.png) 
-
-[Traffic Server Icon, medium](/images/press/tsicon.png) 
-
-[Traffic Server Icon, large](/images/press/tsicon_large.png) 
-
