--- source/markdown/users.mdtext
@@ -1,74 +1,0 @@
-Title: Companies using Apache Traffic Server
-Notice:    Licensed to the Apache Software Foundation (ASF) under one
-           or more contributor license agreements.  See the NOTICE file
-           distributed with this work for additional information
-           regarding copyright ownership.  The ASF licenses this file
-           to you under the Apache License, Version 2.0 (the
-           "License"); you may not use this file except in compliance
-           with the License.  You may obtain a copy of the License at
-           .
-             https://www.apache.org/licenses/LICENSE-2.0
-           .
-           Unless required by applicable law or agreed to in writing,
-           software distributed under the License is distributed on an
-           "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-           KIND, either express or implied.  See the License for the
-           specific language governing permissions and limitations
-           under the License.
-
-		
-
-# Companies using Apache Traffic Server #
-Apache Traffic Server is used world-wide by companies of many sizes, ranging from the 
-small entrepreneurial starter companies to the large Internet conglomerates like Yahoo, 
-Comcast and Akamai. Companies that use Apache Traffic Server include:
-
-<p>
-<a rel="nofollow" href="http://www.comcast.com"><img alt="Comcast" src="/images/companies/comcast.png" title=""/></a>
-<a rel="nofollow" href="http://www.yahoo.com"><img alt="Yahoo!" src="/images/companies/yahoo.png" title=""/></a>
-<a rel="nofollow" href="http://www.linkedin.com/"><img alt="LinkedIn" src="/images/companies/linkedin.png" title=""/></a>
-<a rel="nofollow" href="http://www.godaddy.com/"><img alt="Go Daddy" src="/images/companies/godaddy.png" title=""/></a>
-<a rel="nofollow" href="http://www.yahoo.co.jp/"><img alt="Yahoo! Japan" src="/images/companies/yahoo_japan.png" title=""/></a>
-</p>
-
-<p>
-<a rel="nofollow" href="http://www.akamai.com/"><img alt="Akamai" src="/images/companies/akamai.png" title=""/></a>
-<a rel="nofollow" href="http://www.upyun.com/"><img alt="Upyun" src="/images/companies/upyunLogo.png" title=""/></a>
-<a rel="nofollow" href="http://www.verycloud.cn/"><img alt="Verycloud! China" src="/images/companies/verycloud_logo.png" title=""/></a>
-<a rel="nofollow" href="http://www.websense.com/"><img alt="Websense" src="/images/companies/websense.png" title=""/></a>
-<a rel="nofollow" href="http://www.azion.com"><img alt="Azion Technologies" src="/images/companies/azion.png" title=""/></a>
-</p>
-
-<p>
-<a rel="nofollow" href="http://www.4399.com/"><img alt="4399" src="/images/companies/4399_logo.gif" title=""/></a>
-<a rel="nofollow" href="http://owmobility.com/"><img alt="Openwave Mobility" src="/images/companies/owmobility.png" title=""/></a>
-<a rel="nofollow" href="https://torchbox.com/"><img alt="Torchbox" src="/images/companies/torchbox.png" title=""/></a>
-<a rel="nofollow" href="http://www.iispeed.com/"><img alt="IISpeed" src="/images/companies/iispeed.png" title=""/></a>
-<a rel="nofollow" href="http://www.omniti.com/"><img alt="OmniTI" src="/images/companies/omniti.png" title=""/></a>
-</p>
-
-<p>
-<a rel="nofollow" href="http://www.betfair.com/"><img alt="Betfair" src="/images/companies/betfair.png" title=""/></a>
-<a rel="nofollow" href="http://www.ip-only.com/link=ats"><img alt="IP Only" src="/images/companies/iponly.png" title=""/></a>
-<a rel="nofollow" href="http://www.thelounge.net/"><img alt="The Lounge" src="/images/companies/tthelounge.png" title=""/></a>
-<a rel="nofollow" href="http://www.internode.on.net/"><img alt="Internode" src="/images/companies/internode.png" title=""/></a>
-<a rel="nofollow" href="http://www.sic-software.com/"><img alt="SiC Software" src="/images/companies/sic.png" title=""/></a>
-</p>
-
-<p>
-<a rel="nofollow" href="http://www.devaus.net/"><img alt="Devaus" src="/images/companies/devaus.png" title=""/></a>
-<a rel="nofollow" href="http://www.powerhttp.com/"><img alt="PowerHTTP" src="/images/companies/powerhttp.png" title=""/></a>
-<a rel="nofollow" href="http://vng.com.vn/"><img alt="VNG" src="/images/companies/vng.png" title=""/></a>
-<a rel="nofollow" href="http://www.tilauslentopalvelu.fi/"><img alt="Suomen Tilauslentopalvelu" src="/images/companies/suomen.png" title=""/></a>
-<a rel="nofollow" href="http://www.guruve.com/"><img alt="Guruve" src="/images/companies/guruve.png" title=""/></a>
-</p>
-
-<p>
-<a rel="nofollow" href="http://equalit.ie/"><img alt="eQualit.ie" src="/images/companies/equalit.ie.png" title=""/></a>
-<a rel="nofollow" href="http://frisch-inspiriert.de/"><img alt="Frisch Inspiriert" src="/images/companies/frisch.png" title=""/></a>
-<a rel="nofollow" href="http://www.brainswear.at/"><img alt="Brainswear" src="/images/companies/brainswear.png" title=""/></a>
-<a rel="nofollow" href="http://www.eehu.com/"><img alt="eehu.com" src="/images/companies/eehu.png" title=""/></a>
-<a rel="nofollow" href="http://www.titanbet.co.uk"><img alt="titanbet.co.uk" src="/images/companies/titanbet.png" title=""/></a>
-</p>
-
-Please contact private@trafficserver.apache.org to be added.
