--- source/template.html
@@ -1,91 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "https://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="https://www.w3.org/1999/xhtml"
-  xml:lang="en" lang="en">
-  <head>
-    <title>%TITLE%</title>
-    <!-- Licensed to the Apache Software Foundation (ASF) under one or more contributor license agreements.  See the NOTICE file distributed with this work for additional information regarding copyright ownership.  The ASF licenses this file to you under the Apache License, Version 2.0 (the &quot;License&quot;); you may not use this file except in compliance with the License.  You may obtain a copy of the License at . https://www.apache.org/licenses/LICENSE-2.0 . Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an &quot;AS IS&quot; BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the License for the specific language governing permissions and limitations under the License. -->
-    
-    <!-- The 1140px Grid - https://cssgrid.net/ -->
-    <!-- page_base =  -->
-    <link rel="stylesheet" type="text/css" href="styles/1140.css" media="screen" />
-    <link rel="stylesheet" type="text/css" href="styles/pygments_style.css" />
-    <link rel="stylesheet" type="text/css" href="styles/styles.css" />
-    <link rel="stylesheet" type="text/css" href="https://fonts.googleapis.com/css?family=Droid+Serif:regular,italic,bold" />
-    <style type="text/css">
-        .headerlink, .elementid-permalink {
-            visibility: hidden;
-          }
-    </style>
-    <!--css3-mediaqueries-js - https://code.google.com/p/css3-mediaqueries-js/ - Enables media queries in some unsupported browsers-->
-    <script type="text/javascript" src="js/css3-mediaqueries.js"></script>
-  </head>
-
-  <body class="container">
-    <!-- Header -->
-    <div class="row" id="header">
-      <div class="fourcol">
-        <a href="/"><img class="logo" alt="Apache Traffic Server" src="/images/press/ts_logo_tm_380x73.png" /></a>
-      </div>
-      <div class="threecol">
-        <a href="https://www.apache.org/"><img alt="The Apache Software Foundation" src="https://www.apache.org/images/feather-small.png" /></a>
-      </div>
-      <div class="fivecol right last" id="searchbox">
-        
-        <div id="header-sub">Apache Traffic Server Downloads </div>
-      </div>
-    </div>
-
-    <!-- Navigation (next, previous, up, languages) -->
-    <div class="row nav">
-      
-      <div class="ninecol">
-        <a href="https://trafficserver.apache.org/">Home</a>
-        <a href="/downloads">Downloads</a>
-        <a href="https://docs.trafficserver.apache.org/en/latest/index.html">Documentation</a>
-        <a href="https://cwiki.apache.org/confluence/display/TS/Apache+Traffic+Server">Wiki</a>
-        <a href="https://github.com/apache/trafficserver/issues">Issues</a>
-        <a href="/users.html">Customers</a>
-        <a href="/acknowledgements.html">Acknowledgements</a>
-        <a href="/press.html">Press kit</a>
-        <a href="/assistance.html">Assistance</a>
-      </div>
-      <div class="threecol last right">
-        <a href="https://github.com/apache/trafficserver/security/policy">Security</a>
-        <a href="https://www.apache.org/foundation/sponsorship.html">Donate</a>
-        <a href="https://www.apache.org/foundation/thanks.html">Thanks</a>
-      </div>
-      
-    </div>
-
-    <div class="row content">
-      <div class="twelvecol">
-        <div id="content">
-          %CONTENT%
-        </div>
-      </div>
-      <div class="fourcol last">
-        
-        
-      </div>
-    </div>
-
-    <div id="footer" class="row">
-    <div class="tencol">
-      Copyright  &copy; 2019
-      <a href="https://www.apache.org/">The Apache Software Foundation</a>.
-      Licensed under
-      the <a href="https://www.apache.org/licenses/">Apache License</a>,
-      Version 2.0. Apache Traffic Server, Apache,
-      the Apache Traffic Server logo, and the Apache feather logo are
-      trademarks of The Apache Software Foundation.
-      </div>
-      <div class="twocol right last">
-      <span id="apache_logo">
-        <a href="https://www.apache.org/"><img alt="The Apache Software Foundation" src="https://www.apache.org/images/feather-small.png" /></a>
-      </span>
-      </div>
-    </div>
-
-  </body>
-</html>
