+++ source/templates/downloads.template.html
@@ -29,10 +29,10 @@ To update versions:
          See the License for the specific language governing permissions and limitations under the License. -->
     
     <!-- Tailwind CSS -->
-    <link rel="stylesheet" href="/newsite/styles/tailwind-complete.css">
+    <link rel="stylesheet" href="/styles/tailwind-complete.css">
     
     <!-- Custom Styles -->
-    <link rel="stylesheet" href="/newsite/styles/custom.css">
+    <link rel="stylesheet" href="/styles/custom.css">
 </head>
 <body class="font-sans text-gray-800 bg-white">
     <!-- Header -->
@@ -40,18 +40,18 @@ To update versions:
         <div class="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
             <div class="flex justify-between items-center py-4">
                 <div class="flex items-center">
-                    <a href="/newsite/" class="flex items-center space-x-3">
+                    <a href="/" class="flex items-center space-x-3">
                         <img src="/images/press/ts_logo_tm_380x73.png" alt="Apache Traffic Server" class="h-12 w-auto">
                     </a>
                 </div>
                 
                 <nav class="hidden lg:flex items-center space-x-6">
-                    <a href="/newsite/" class="text-gray-700 hover:text-accent transition-colors font-medium">Home</a>
-                    <a href="/newsite/downloads.html" class="text-accent font-semibold">Downloads</a>
+                    <a href="/" class="text-gray-700 hover:text-accent transition-colors font-medium">Home</a>
+                    <a href="/downloads.html" class="text-accent font-semibold">Downloads</a>
                     <a href="https://docs.trafficserver.apache.org/en/latest/index.html" class="text-gray-700 hover:text-accent transition-colors font-medium">Documentation</a>
                     <a href="https://github.com/apache/trafficserver/issues" class="text-gray-700 hover:text-accent transition-colors font-medium">Issues</a>
                     <a href="https://ci.trafficserver.apache.org/" class="text-gray-700 hover:text-accent transition-colors font-medium">CI</a>
-                    <a href="/newsite/users.html" class="text-gray-700 hover:text-accent transition-colors font-medium">Customers</a>
+                    <a href="/users.html" class="text-gray-700 hover:text-accent transition-colors font-medium">Customers</a>
                 </nav>
                 
                 <div class="hidden lg:flex items-center space-x-4">
@@ -70,12 +70,12 @@ To update versions:
             
             <nav id="mobile-menu" class="hidden lg:hidden pb-4">
                 <div class="space-y-2">
-                    <a href="/newsite/" class="block px-3 py-2 rounded-md text-gray-700 hover:bg-gray-100 hover:text-accent transition-colors font-medium">Home</a>
-                    <a href="/newsite/downloads.html" class="block px-3 py-2 rounded-md bg-blue-50 text-accent font-semibold">Downloads</a>
+                    <a href="/" class="block px-3 py-2 rounded-md text-gray-700 hover:bg-gray-100 hover:text-accent transition-colors font-medium">Home</a>
+                    <a href="/downloads.html" class="block px-3 py-2 rounded-md bg-blue-50 text-accent font-semibold">Downloads</a>
                     <a href="https://docs.trafficserver.apache.org/en/latest/index.html" class="block px-3 py-2 rounded-md text-gray-700 hover:bg-gray-100 hover:text-accent transition-colors font-medium">Documentation</a>
                     <a href="https://github.com/apache/trafficserver/issues" class="block px-3 py-2 rounded-md text-gray-700 hover:bg-gray-100 hover:text-accent transition-colors font-medium">Issues</a>
                     <a href="https://ci.trafficserver.apache.org/" class="block px-3 py-2 rounded-md text-gray-700 hover:bg-gray-100 hover:text-accent transition-colors font-medium">CI</a>
-                    <a href="/newsite/users.html" class="block px-3 py-2 rounded-md text-gray-700 hover:bg-gray-100 hover:text-accent transition-colors font-medium">Customers</a>
+                    <a href="/users.html" class="block px-3 py-2 rounded-md text-gray-700 hover:bg-gray-100 hover:text-accent transition-colors font-medium">Customers</a>
                 </div>
             </nav>
         </div>
@@ -266,8 +266,8 @@ To update versions:
                 <div>
                     <h3 class="text-sm font-semibold text-gray-900 uppercase tracking-wider mb-4">Community</h3>
                     <ul class="space-y-3">
-                        <li><a href="/newsite/users.html" class="text-secondary hover:text-accent transition-colors">Customers</a></li>
-                        <li><a href="/newsite/assistance.html" class="text-secondary hover:text-accent transition-colors">Get Help</a></li>
+                        <li><a href="/users.html" class="text-secondary hover:text-accent transition-colors">Customers</a></li>
+                        <li><a href="/assistance.html" class="text-secondary hover:text-accent transition-colors">Get Help</a></li>
                         <li><a href="https://github.com/apache/trafficserver/issues" class="text-secondary hover:text-accent transition-colors">Issues</a></li>
                         <li><a href="https://the-asf.slack.com/" class="text-secondary hover:text-accent transition-colors">Slack Channel</a></li>
                     </ul>
@@ -287,8 +287,8 @@ To update versions:
                 <div>
                     <h3 class="text-sm font-semibold text-gray-900 uppercase tracking-wider mb-4">More</h3>
                     <ul class="space-y-3">
-                        <li><a href="/newsite/press.html" class="text-secondary hover:text-accent transition-colors">Press Kit</a></li>
-                        <li><a href="/newsite/acknowledgements.html" class="text-secondary hover:text-accent transition-colors">Acknowledgements</a></li>
+                        <li><a href="/press.html" class="text-secondary hover:text-accent transition-colors">Press Kit</a></li>
+                        <li><a href="/acknowledgements.html" class="text-secondary hover:text-accent transition-colors">Acknowledgements</a></li>
                         <li><a href="https://x.com/trafficserver" class="text-secondary hover:text-accent transition-colors">X</a></li>
                     </ul>
                 </div>
@@ -310,7 +310,7 @@ To update versions:
     </footer>
     
     <!-- Mobile Menu Script -->
-    <script src="/newsite/js/menu.js"></script>
+    <script src="/js/menu.js"></script>
 </body>
 </html>
